Add styling. Add delete task confirm box

+++ server/public/scripts/client.js
@@ -7,8 +7,8 @@ console.log('client.js ready, jquery ready');
     $('#submitBtn').on('click',addTask);
     $('#tbody').on('click', '.delBtn', deleteTask);
     $('#tbody').on('click','.check',changeStatus);
+    /*$('#tbody').on('mouseout','.noteBox',changeNotes);*/
  }
-
 function showTasks() {
      console.log("in showTasks");
     $.ajax({
@@ -29,15 +29,18 @@ function render(taskObj) {
         
         const checkClass = x.completed ? "checkClass" : "";
         const checkText = x.completed ? "" : "";
-            
+        
+        const textClass = x.completed ? "textClass" : "";
+ 
             $('#tbody').append(`
             <tr class=${tasksClass}>
-                <td> ${x.task}</td> 
+                <td> - ${x.task}</td> 
                 <td> ${x.completed} </td>
                 <td class=${checkClass}> <input type="checkbox" data-id="${x.id}" class="check" ${checkText}> ${btnText} </input></td>
-                <td> ${x.notes} </td>
+                <td class=${textClass}><textarea class="noteBox" data-note="${x.notes}">${x.notes} </textarea></td>
                 <td> <button class="delBtn" data-id="${x.id}"> Delete </button></td>
              </tr>
+              
                `)
             }
 }
@@ -47,10 +50,10 @@ function addTask(evt) {
 if($('#inputTask').val() == "") {
     alert("Please enter a task");
     return;
-}
-console.log("Submit button clicked");
+} 
 
-    let taskObj = {
+console.log("Submit button clicked");
+     let taskObj = {
         task: $('#inputTask').val(),
         completed: false,
         notes: $('#inputNotes').val()
@@ -69,18 +72,23 @@ console.log("Submit button clicked");
 
 function deleteTask(evt) {
     evt.preventDefault();
-console.log('DELETE button was clicked,ID being', $(this).data('id'));
+    console.log('DELETE button was clicked,ID being', $(this).data('id'));
     let taskId = $(this).data('id');
- 
-    $.ajax({
-        method: 'DELETE',
-        url: `/tasks/${taskId}`,
-    }).then(function(response) {
-        console.log("The song was deleted!");
-        showTasks();
-    }).catch(function(error) {
-        console.log("ERROR on delete",error);
-    });
+    let message = "Are you sure you want to delete this task?";
+    if (confirm(message) == true) {
+        $.ajax({
+            method: 'DELETE',
+            url: `/tasks/${taskId}`,
+        }).then(function(response) {
+            console.log("The song was deleted!");
+            showTasks();
+        }).catch(function(error) {
+            console.log("ERROR on delete",error);
+        });
+    } else {
+      return;
+    }
+
 };
 
 function changeStatus(evt) {
@@ -101,3 +109,28 @@ function changeStatus(evt) {
     });
 };
 
+/*
+function changeNotes(evt) {
+     evt.preventDefault();
+    let noteId = $(this).data('note');
+    console.log(noteId);
+
+    let noteObj = {
+        newNotes:  $(this).data('note')
+    }
+     $.ajax({
+        method: 'PUT',
+        url: `/tasks/${noteId}`,
+        data:noteObj
+     })
+    .then(function(response) {
+        console.log("The note was updated");
+        console.log(response,"in .then notes");
+        showTasks();
+    })
+    .catch(function(err) {
+        console.log("Err on notebox",err);
+    });
+};
+
+*/
